Always use UnsafeRow for initial agg buffer.
